Switched Tuple, Point and Vector to use double instead of float. Added Vector normalization and Vector magnitude operators.

+++ RayTracerChallenge/Point.cs
@@ -2,6 +2,6 @@
 
 public class Point : Tuple
 {
-    public Point(float x, float y, float z) : base(x,y,z,1) {}
+    public Point(double x, double y, double z) : base(x,y,z,1) {}
 }
